Remove mocha

+++ test/unit/status_test.rb
@@ -19,83 +19,95 @@ class StatusTest < ActiveSupport::TestCase
   end
 
   def test_enabled_is_false_when_switch_file_exists
-    Rails.expects(:root).at_most(2).returns(@test_root)
-    create_switch_file
-    assert_equal false, Flog::Status.enabled?
+    Rails.stub(:root, @test_root) do
+      create_switch_file
+      assert_equal false, Flog::Status.enabled?
+    end
   end
 
   def test_enabled_is_true_when_error_is_raised_in_process
-    Rails.expects(:root).returns(nil) # For raise NoMethodError
-    create_switch_file
-    assert Flog::Status.enabled?
+    Rails.stub(:root, nil) do # For raise NoMethodError
+      create_switch_file
+      assert Flog::Status.enabled?
+    end
   end
 
   def test_sql_formattable_is_true_when_enable_and_sql_switch_file_does_not_exist
-    Rails.expects(:root).at_most(2).returns(@test_root)
-    delete_switch_file
-    delete_sql_switch_file
-    assert Flog::Status.sql_formattable?
+    Rails.stub(:root, @test_root) do
+      delete_switch_file
+      delete_sql_switch_file
+      assert Flog::Status.sql_formattable?
+    end
   end
 
   def test_sql_formattable_is_false_when_enable_and_sql_switch_file_exists
-    Rails.expects(:root).at_most(2).returns(@test_root)
-    delete_switch_file
-    create_sql_switch_file
-    assert_equal false, Flog::Status.sql_formattable?
+    Rails.stub(:root, @test_root) do
+      delete_switch_file
+      create_sql_switch_file
+      assert_equal false, Flog::Status.sql_formattable?
+    end
   end
 
   def test_sql_formattable_is_false_when_disable_and_sql_switch_file_not_exist
-    Rails.expects(:root).at_most(2).returns(@test_root)
-    create_switch_file
-    delete_sql_switch_file
-    assert_equal false, Flog::Status.sql_formattable?
+    Rails.stub(:root, @test_root) do
+      create_switch_file
+      delete_sql_switch_file
+      assert_equal false, Flog::Status.sql_formattable?
+    end
   end
 
   def test_sql_formattable_is_true_when_error_is_raised_in_process
-    Rails.expects(:root).at_most(2).returns(nil) # For raise NoMethodError
-    create_sql_switch_file
-    assert Flog::Status.sql_formattable?
+    Rails.stub(:root, nil) do # For raise NoMethodError
+      create_sql_switch_file
+      assert Flog::Status.sql_formattable?
+    end
   end
 
   def test_sql_formattable_is_false_when_disable_and_sql_switch_file_exists
-    Rails.expects(:root).at_most(2).returns(@test_root)
-    create_switch_file
-    create_sql_switch_file
-    assert_equal false, Flog::Status.sql_formattable?
+    Rails.stub(:root, @test_root) do
+      create_switch_file
+      create_sql_switch_file
+      assert_equal false, Flog::Status.sql_formattable?
+    end
   end
 
   def test_params_formattable_is_true_when_enable_and_params_switch_file_does_not_exist
-    Rails.expects(:root).at_most(2).returns(@test_root)
-    delete_switch_file
-    delete_params_switch_file
-    assert Flog::Status.params_formattable?
+    Rails.stub(:root, @test_root) do
+      delete_switch_file
+      delete_params_switch_file
+      assert Flog::Status.params_formattable?
+    end
   end
 
   def test_params_formattable_is_false_when_enable_and_params_switch_file_exists
-    Rails.expects(:root).at_most(2).returns(@test_root)
-    delete_switch_file
-    create_params_switch_file
-    assert_equal false, Flog::Status.params_formattable?
+    Rails.stub(:root, @test_root) do
+      delete_switch_file
+      create_params_switch_file
+      assert_equal false, Flog::Status.params_formattable?
+    end
   end
 
   def test_params_formattable_is_false_when_disable_and_params_switch_file_not_exist
-    Rails.expects(:root).at_most(2).returns(@test_root)
-    create_switch_file
-    delete_params_switch_file
-    assert_equal false, Flog::Status.params_formattable?
+    Rails.stub(:root, @test_root) do
+      create_switch_file
+      delete_params_switch_file
+      assert_equal false, Flog::Status.params_formattable?
+    end
   end
 
   def test_params_formattable_is_false_when_disable_and_params_switch_file_exists
-    Rails.expects(:root).at_most(2).returns(@test_root)
-    create_switch_file
-    create_params_switch_file
-    assert_equal false, Flog::Status.params_formattable?
+    Rails.stub(:root, @test_root) do
+      create_switch_file
+      create_params_switch_file
+      assert_equal false, Flog::Status.params_formattable?
+    end
   end
 
   def test_params_formattable_is_true_when_error_is_raised_in_process
-    Rails.expects(:root).at_most(2).returns(nil) # For raise NoMethodError
-    create_params_switch_file
-    assert Flog::Status.params_formattable?
+    Rails.stub(:root, nil) do # For raise NoMethodError
+      create_params_switch_file
+      assert Flog::Status.params_formattable?
+    end
   end
 
   private
@@ -128,8 +140,6 @@ class StatusTest < ActiveSupport::TestCase
   end
 
   def delete_file(file_path)
-    if File.exist?(file_path)
-      File.delete(file_path)
-    end
+    File.delete(file_path) if File.exist?(file_path)
   end
 end
